16407007 Folder restructuring (#28)
