+++ .eslintrc
@@ -26,7 +26,7 @@
     "jsdoc/check-param-names": 1,
     "jsdoc/check-tag-names": 1,
     "jsdoc/check-types": 1,
-    "jsdoc/newline-after-description": false,
+    "jsdoc/newline-after-description": 0,
     "jsdoc/no-undefined-types": 1,
     "jsdoc/require-description": 1,
     "jsdoc/require-description-complete-sentence": 1,
