+++ controllers/articlesController.js
@@ -1,6 +1,6 @@
 import models from "../models";
 import constants from "../helpers/constants";
-import articleValidator from "../helpers/validators/article";
+import articleValidator from "../helpers/validators/articleValidators";
 
 const { User, Article, Tag, ArticleTag } = models;
 const { CREATED, NOT_FOUND, BAD_REQUEST } = constants.statusCode;
@@ -22,7 +22,7 @@ export default class PostController {
       if (user && valid) {
         const article = await Article.create({ ...rest, userId });
         const { id: articleId } = article.dataValues;
-        for (let tag of tagsList) {
+        for (const tag of tagsList) {
           const tags = await Tag.findOrCreate({ where: { name: tag } });
           refs.push(
             await ArticleTag.create({
@@ -46,7 +46,7 @@ export default class PostController {
         return res
           .status(BAD_REQUEST)
           .send({ message: error.details[0].message, status: BAD_REQUEST });
-      else return res.status(500).send({ message: error, status: 500 });
+      return res.status(500).send({ message: error, status: 500 });
     }
   }
 }
