+++ controllers/profilesController.js
@@ -1,6 +1,7 @@
 import models from "../models";
-import validation from "../middlewares/validators/updateProfileValidator";
-const { User, Follower } = models;
+import validation from "../middlewares/updateProfileValidator";
+
+const { User } = models;
 
 /**
  * Profile Class
@@ -10,29 +11,29 @@ const { User, Follower } = models;
 class Profile {
   /**
    *
-   * Update profile
+   * Update profile.
    *
-   * @param  {object} req - The request object
-   * @param  {object} res - The response object
-   * @return {object} - It returns the response object
+   * @param  {Object} req - The request object.
+   * @param  {Object} res - The response object.
+   * @return {Object} - It returns the response object.
    */
   static async update(req, res) {
-    let usernameParameter = req.params.username;
-    let usernameFromtoken = req.user.username;
-    let newUser = validation(req.body.profile);
+    const usernameParameter = req.params.username;
+    const usernameFromToken = req.user.username;
+    const newUser = validation(req.body.profile);
     if (newUser.error) {
       return res.status(400).json({ error: newUser.error });
     }
-    if (usernameFromtoken != usernameParameter) {
+    if (usernameFromToken !== usernameParameter) {
       return res.status(403).json({ error: "Not authorized" });
     }
 
     try {
-      let updatedProfile = await User.update(newUser, {
+      const updatedProfile = await User.update(newUser, {
         returning: true,
-        where: { username: usernameFromtoken }
+        where: { username: usernameFromToken }
       });
-      let newProfile = updatedProfile[1][0];
+      const newProfile = updatedProfile[1][0];
       return res
         .status(200)
         .json({ message: "Profile updated successfully", profile: newProfile });
@@ -43,14 +44,14 @@ class Profile {
 
   /**
    *
-   * getProfile
+   * GetProfile.
    *
-   * @param  {object} req - The request object
-   * @param  {object} res - The response object
-   * @return {object} - It returns the response object
+   * @param  {Object} req - The request object.
+   * @param  {Object} res - The response object.
+   * @return {Object} - It returns the response object.
    */
   static async getProfile(req, res) {
-    let { username } = req.params;
+    const { username } = req.params;
     try {
       const profile = await User.findOne({
         where: { username }
@@ -60,7 +61,7 @@ class Profile {
       }
       return res.status(200).json({ profile });
     } catch (error) {
-      if (error.message == "No user with that name") {
+      if (error.message === "No user with that name") {
         return res.status(400).json({ message: "No user with that name" });
       }
       return res.status(500).json({ message: "Error happened", error });
@@ -69,11 +70,11 @@ class Profile {
 
   /**
    *
-   * getAllProfiles
+   * GetAllProfiles.
    *
-   * @param  {object} req - The request object
-   * @param  {object} res - The response object
-   * @return {object} - It returns the response object
+   * @param  {Object} req - The request object.
+   * @param  {Object} res - The response object.
+   * @return {Object} - It returns the response object.
    */
   static async getAllProfiles(req, res) {
     try {
@@ -86,28 +87,28 @@ class Profile {
 
   /**
    *
-   * delete
+   * Delete.
    *
-   * @param  {object} req - The request object
-   * @param  {object} res - The response object
-   * @return {object} - It returns the response object
+   * @param  {Object} req - The request object.
+   * @param  {Object} res - The response object.
+   * @return {Object} - It returns the response object.
    */
   static async delete(req, res) {
-    let { username } = req.params;
-    let usernameFromToken = req.user.username;
-    if (username != usernameFromToken) {
+    const { username } = req.params;
+    const usernameFromToken = req.user.username;
+    if (username !== usernameFromToken) {
       return res.status(403).json({ message: "Unauthorized request" });
     }
     try {
       const deletedUser = await User.destroy({
         where: { username }
       });
-      if (deletedUser == 0) {
+      if (deletedUser === 0) {
         throw new Error("There no user with that username");
       }
       res.status(200).json({ message: "Profile deleted successfully" });
     } catch (error) {
-      if ((error.message = "There no user with that username")) {
+      if (error.message === "There no user with that username") {
         return res.status(400).json({ message: `There no user with that username` });
       }
       res.status(500).json({ message: "Error happened", error });
