--- helpers/data.js
@@ -1,12 +1,0 @@
-export const data = {
-  dummyUser: {
-    email: "luc.bayo@gmail.com",
-    password: "password",
-    username: "luc2017"
-  },
-  dummyUser2: {
-    email: "fake.email@gmail.com",
-    password: "password",
-    username: "jean786786"
-  }
-};
