+++ routes/profile.js
@@ -1,40 +1,25 @@
 import Router from "express";
-import passport from "passport";
-import Profile from "../controllers/profileController";
-import FollowerController from "../controllers/followerController";
+import Profile from "../controllers/profilesController";
+import FollowerController from "../controllers/followersController";
+import checkAuth from "../middlewares/checkAuth";
+
 const profileRouter = Router();
 
 profileRouter
-  .put(
-    "/profiles/:username",
-    passport.authenticate("jwt", { session: false }),
-    Profile.update
-  )
+  .put("/profiles/:username", checkAuth, Profile.update)
   .get("/profiles/:username", Profile.getProfile)
   .get("/profiles", Profile.getAllProfiles)
-  .delete(
-    "/profiles/:username",
-    passport.authenticate("jwt", { session: false }),
-    Profile.delete
-  )
-  .post(
-    "/profiles/:username/follow",
-    passport.authenticate("jwt", { session: false }),
-    FollowerController.followUser
-  )
-  .delete(
-    "/profiles/:username/follow",
-    passport.authenticate("jwt", { session: false }),
-    FollowerController.unFollow
-  )
+  .delete("/profiles/:username", checkAuth, Profile.delete)
+  .post("/profiles/:username/follow", checkAuth, FollowerController.followUser)
+  .delete("/profiles/:username/follow", checkAuth, FollowerController.unFollow)
   .get(
     "/profiles/:username/followers",
-    passport.authenticate("jwt", { session: false }),
+    checkAuth,
     FollowerController.getAllFollowers
   )
   .get(
     "/profiles/:username/followings",
-    passport.authenticate("jwt", { session: false }),
+    checkAuth,
     FollowerController.getFollowings
   );
